pyproject.toml
[project]
name = "cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai[tools]>=0.108.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.0",
    "streamlit-javascript>=0.1.5",
]
